+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 * bug-fix: Estimators: Fix attach for LDA
 * bug-fix: Estimators: allow code_location to have no key prefix
+* bug-fix: Local Mode: Fix s3 training data download when there is a trailing slash
 * feature: Allow Local Serving of Models in S3